lesson 3-4

+++ academy8121/src/by/academy/lesson3/operators/CharInc.java
@@ -1,9 +1,10 @@
 package by.academy.lesson3.operators;
+
 public class CharInc {
-    public static void main(String[] args) {
-        char ch = 'X';
-        System.out.println("ch содержит " + ch);
-        ch++; // увеличиваем ch
-        System.out.println("Новое значение ch: " + ch);
-    }
+	public static void main(String[] args) {
+		char ch = 'X';
+		System.out.println("ch содержит " + (int) ch);
+		ch++; // увеличиваем ch
+		System.out.println("Новое значение ch: " + ch);
+	}
 }
